services/storage_account/func_get_last_fetch_time_for_each_storage_account/requirements.txt
azure-functions
pytest==7.4.2
black==23.11.0
python-dotenv==1.0.0
pytest-cov
adal
requests